Split SignRPMsKojiThread into two
This fixes a deadlock in sign-rpms where KojiThread blocks on
SendRequestThread which blocks on the server which blocks on
ReadReplyThread which blocks on KojiThread.

IIRC there is a single KojiThread only to minimize load on Koji, which
is not as much a concern as completing; this will get us to two Koji
requests at a time at most.