Fix rpm signing with Fedora vault
This fixes some oddities in the F25 rpmsign with gnupg1 combination.
For some reason, it now sometimes asks for a passphrase twice before
accepting it.
The output text has also slightly changed.
In addition to that, the newer rpm/gpg combinations no longer work
with --batch --passphrase-fd 3, whereas the older version still
require those. For that case, we now detect whether they are specified
in the original gpg_sign_cmd and insert them if needed.

Signed-off-by: Patrick Uiterwijk <puiterwijk@redhat.com>