Implement Head Signing
This adds an option to the client to request Head Signing.
Head Signing will only sign the RPM Header (v4 signature scheme).
This saves significant amounts of data that don't have to be
transferred, because the signature is computed over just the
header.

Note: The koji paths of this patch have been tested minimally.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>