Increase RPM size for testing head signing
It turns out that we were accidentally still calling rpm verify
with RPM head signing, which has no chance of working if not
the full RPM was sent (as in head signing).
This wasn't triggered by the test suite because the RPMs
that were used to test were too small.

This commit makes sure the test suite generates RPMs
that are big enough to be impacted by header-only sending.
They generate 5MB data per RPM, in order to be big enough.
Random data is generated to defeat the compression done
by RPM, to ensure the resulting RPM is actually 5MB.

Signed-off-by: Patrick Uiterwijk <patrick@puiterwijk.org>